feat: export tree path for new intent ANCESTOR_MIGRATED
